src/memory/meta_learner.py
import numpy as np
from sklearn.linear_model import LinearRegression
from sklearn.preprocessing import StandardScaler


class MetaLearner:
    def __init__(self):
        self.model = LinearRegression()
        self.scaler = StandardScaler()
        self.is_fitted = False
    
    def partial_fit(self, X, y):
        # Update model with new data
        if not self.is_fitted:
            X_scaled = self.scaler.fit_transform(X)
            self.model.fit(X_scaled, y)
            self.is_fitted = True
        else:
            X_scaled = self.scaler.transform(X)
            self.model.partial_fit(X_scaled, y)
    
    def predict(self, repo_type):
        # Predict best approach for a given repo type
        try:
            # This would be more complex in practice
            return "recommended_approach"
        except Exception as e:
            print(f"[ERROR] Failed to predict: {str(e)}")
            return "default_approach"